aggiunti e sistemati controlli sui database
